SaberLockSys: Replace with limited RPS system
Ref c84e3f2350a12c077a121ed923f0ef1f06f5ce00,
which added a movement based "guessing" system.
Limited RPS is a quicker and perhaps simpler variant of that.

In Limited RPS system: there are rock, paper and scissor cards.
Each player starts with 2 of each. In a saberlock you both chose 1.
Then we resolve like in normal RPS, but the card is used up.
On respawn you get a random card (max 2 of each).
On sabering someone, you get a scissors card.
If you have no cards, you automatically lose.
If both have no cards, the saberlock is "cancelled".
There is room for a bit of strategy here, but needs playtesting.